% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglyr.R
\name{gg_bound_sf}
\alias{gg_bound_sf}
\title{Set ggplot extent to SF object}
\usage{
gg_bound_sf(sf_lyr)
}
\arguments{
\item{sf_lyr}{Simple Features layer to set extent.}
}
\value{
coord_sf ggplot object
}
\description{
Sets the extent of a ggplot to the extent of a given Simple Features object.
}
