% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglyr.R
\name{gglyr_save}
\alias{gglyr_save}
\title{Save gglyr}
\usage{
gglyr_save(lyr, name, comment, overwrite = T)
}
\arguments{
\item{lyr}{ggproto or function object to save}

\item{name}{A name for the layer - this should be short, ideally snakecase,
and will be used in code to retrieve the layer.}

\item{comment}{A longer comment to be saved with the layer.}

\item{overwrite}{Overwrite current layer if it exists? Defaults to true.}
}
\value{
No value
}
\description{
Saves a ggproto object to the currently set gglyr library.
}
