% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglyr.R
\name{gglyr_set_lib}
\alias{gglyr_set_lib}
\title{Set gglyr library path}
\usage{
gglyr_set_lib(lib_path)
}
\arguments{
\item{lib_path}{Path to gglyr library}
}
\value{
No value
}
\description{
Sets the gglyr library path to a local directory.
}
